# app/http/swagger/swagger.yaml
basePath: /
definitions:
  demo.UserDTO:
    properties:
      id:
        type: integer
      name:
        type: string
    type: object
  user.loginParam:
    properties:
      password:
        type: string
      username:
        type: string
    required:
    - password
    - username
    type: object
  user.registerParam:
    properties:
      email:
        type: string
      password:
        type: string
      username:
        type: string
    required:
    - email
    - password
    - username
    type: object
info:
  contact:
    email: fastopencn
    name: fastopencn1
  description: Colon测试
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: https://github.com/swaggo/swag
  title: Colon
  version: "1.1"
paths:
  /demo/demo:
    get:
      description: 获取所有用户
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              items:
                $ref: '#/definitions/demo.UserDTO'
              type: array
            type: array
      summary: 获取所有用户
      tags:
      - demo
  /demo/demo2:
    get:
      description: 获取所有学生,不进行分页
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/demo.UserDTO'
            type: array
      summary: 获取所有学生
      tags:
      - demo
  /demo/hello:
    get:
      description: hello
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              items:
                $ref: '#/definitions/demo.UserDTO'
              type: array
            type: array
      summary: hello
      tags:
      - demo
  /user/login:
    post:
      consumes:
      - application/json
      description: 用户登录接口
      parameters:
      - description: login with param
        in: body
        name: loginParam
        required: true
        schema:
          $ref: '#/definitions/user.loginParam'
      produces:
      - application/json
      responses:
        "200":
          description: token
          schema:
            type: string
      summary: 用户登录
      tags:
      - user
  /user/logout:
    get:
      consumes:
      - application/json
      description: 注销用户登录信息
      produces:
      - application/json
      responses:
        "200":
          description: 用户退出成功
          schema:
            type: string
      summary: 用户登出
      tags:
      - user
  /user/register:
    post:
      consumes:
      - application/json
      description: 用户注册接口
      parameters:
      - description: 注册参数
        in: body
        name: registerParam
        required: true
        schema:
          $ref: '#/definitions/user.registerParam'
      produces:
      - application/json
      responses:
        "200":
          description: 注册成功
          schema:
            type: string
      summary: 用户注册
      tags:
      - user
  /user/register/verify:
    get:
      consumes:
      - application/json
      description: 使用token验证用户注册信息
      parameters:
      - description: 注册token
        in: query
        name: token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 注册成功，请进入登录页面
          schema:
            type: string
      summary: 验证注册信息
      tags:
      - user
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
  BasicAuth:
    type: basic
swagger: "2.0"
x-extension-openapi:
  example: value on a json format